requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
gspread==5.7.2
oauth2client==4.1.3
pytz==2023.3
requests==2.31.0
gspread-formatting==1.2.1
Flask==2.3.3
pytest
pytest-asyncio
asynctest